+++ cmsdev/internal/test/bos/bos_session.go
@@ -33,105 +33,223 @@ package bos
 import (
 	"net/http"
 	"stash.us.cray.com/SCMS/cms-tools/cmsdev/internal/lib/common"
-	"stash.us.cray.com/SCMS/cms-tools/cmsdev/internal/lib/test"
 )
 
-// session tests
-func sessionTestsAPI() bool {
-	var baseurl string = common.BASEURL
-	const totalNumTests int = 1
+const bosV1SessionsUri = bosV1BaseUri + "/session"
+const bosV2SessionsUri = bosV2BaseUri + "/sessions"
 
-	numTests, numTestsFailed := 0, 0
-	params := test.GetAccessTokenParams()
-	if params == nil {
+const bosV1SessionsCLI = "session"
+const bosV2SessionsCLI = "sessions"
+const bosDefaultSessionsCLI = bosV2SessionsCLI
+
+func sessionsTestsAPI(params *common.Params) (passed bool) {
+	passed = true
+
+	// v1 sessions
+	if !sessionsV1TestsURI(bosV1SessionsUri, params) {
+		passed = false
+	}
+
+	// v2 sessions
+	if !sessionsV2TestsURI(bosV2SessionsUri, params) {
+		passed = false
+	}
+
+	return
+}
+
+func sessionsTestsCLI() (passed bool) {
+	passed = true
+
+	// v1 sessions
+	if !sessionsV1TestsCLICommand("v1", bosV1SessionsCLI) {
+		passed = false
+	}
+
+	// v2 sessions
+	if !sessionsV2TestsCLICommand("v2", bosV2SessionsCLI) {
+		passed = false
+	}
+
+	// default (v2) sessions
+	if !sessionsV2TestsCLICommand(bosDefaultSessionsCLI) {
+		passed = false
+	}
+
+	return
+}
+
+// v1 sessions API tests
+func sessionsV1TestsURI(uri string, params *common.Params) bool {
+	// test #1, list session
+	common.Infof("GET %s test scenario", uri)
+	resp, err := bosRestfulVerifyStatus("GET", uri, params, http.StatusOK)
+	if err != nil {
+		common.Error(err)
 		return false
 	}
 
-	// test #1, list session
-	url := baseurl + endpoints["bos"]["session"].Url
-	numTests++
-	test.RestfulTestHeader("GET session", numTests, totalNumTests)
-	resp, err := test.RestfulVerifyStatus("GET", url, *params, http.StatusOK)
+	// BOS v1: Validate that at least we can decode the JSON into a list of strings
+	stringList, err := common.DecodeJSONIntoStringList(resp.Body())
 	if err != nil {
 		common.Error(err)
-		numTestsFailed++
-	} else {
-		// Validate that at least we can decode the JSON into a list of strings
-		stringList, err := common.DecodeJSONIntoStringList(resp.Body())
-		if err != nil {
-			common.Error(err)
-			numTestsFailed++
-		} else if len(stringList) == 0 {
-			common.VerbosePrintDivider()
-			common.Infof("skipping test GET /session/{session_id}")
-			common.Infof("results from previous test is []")
-		} else {
-			// a session_id is available
-			sessionId := stringList[0]
-
-			url = baseurl + endpoints["bos"]["session"].Url + "/" + sessionId
-			numTests++
-			test.RestfulTestHeader("GET session_id", numTests, totalNumTests)
-			resp, err = test.RestfulVerifyStatus("GET", url, *params, http.StatusOK)
-			if err != nil {
-				common.Error(err)
-				numTestsFailed++
-			} else {
-				// Validate that the session can at least be decoded into a string map
-				_, err = common.DecodeJSONIntoStringMap(resp.Body())
-				if err != nil {
-					common.Error(err)
-					numTestsFailed++
-				}
-			}
-		}
-		// TODO: deeper validation of returned response
+		return false
+	} else if len(stringList) == 0 {
+		common.Infof("skipping test GET %s/{session_id}", uri)
+		common.Infof("results from previous test is []")
+		return true
 	}
 
-	test.RestfulTestResultSummary(numTestsFailed, numTests)
-	return numTestsFailed == 0
+	// a session_id is available
+	sessionId := stringList[0]
+	common.Infof("Found BOS v1 session with ID '%s'", sessionId)
+
+	// test #2, describe session with session_id
+	uri += "/" + sessionId
+	if !basicGetUriVerifyStringMapTest(uri, params) {
+		return false
+	}
+
+	return true
 }
 
-func sessionTestsCLI(vnum int) bool {
-	// test #1, list session
-	common.Infof("Getting list of all BOS sessions via CLI")
-	cmdOut := runCLICommand(vnum, "session", "list")
+func ValidateV2Session(mapCmdOut []byte, expectedName string) bool {
+	// For BOSv2, the session object we get back should have a 'name' field matching the name that we requested.
+	// So let's validate that
+	err := common.ValidateStringFieldValue("BOS session", "name", expectedName, mapCmdOut)
+	if err != nil {
+		common.Error(err)
+		return false
+	}
+	return true
+}
+
+// v2 sessions API tests
+func sessionsV2TestsURI(uri string, params *common.Params) bool {
+	// test #1, list sessions
+	common.Infof("GET %s test scenario", uri)
+	resp, err := bosRestfulVerifyStatus("GET", uri, params, http.StatusOK)
+	if err != nil {
+		common.Error(err)
+		return false
+	}
+
+	// BOS v2: Decode JSON into a list
+	sessionList, err := common.DecodeJSONIntoList(resp.Body())
+	if err != nil {
+		common.Error(err)
+		return false
+	} else if len(sessionList) == 0 {
+		common.Infof("skipping test GET %s/{session_id}", uri)
+		common.Infof("results from previous test is []")
+		return true
+	}
+
+	// Take the first session listed. It should be a dictionary object.
+	sessionObject := sessionList[0]
+	session, ok := sessionObject.(map[string]interface{})
+	if !ok {
+		common.Errorf("First BOS session listed is not a dictionary object: %v", sessionObject)
+		return false
+	}
+
+	common.Debugf("Getting 'name' field BOS session: %v", session)
+	sessionId, err := common.GetStringFieldFromMapObject("name", session)
+	if err != nil {
+		common.Error(err)
+		return false
+	}
+	common.Infof("Found BOS v2 session with name '%s'", sessionId)
+
+	// test #2: describe session
+	uri += "/" + sessionId
+	resp, err = bosRestfulVerifyStatus("GET", uri, params, http.StatusOK)
+	if err != nil {
+		common.Error(err)
+		return false
+	} else if !ValidateV2Session(resp.Body(), sessionId) {
+		return false
+	}
+
+	return true
+}
+
+// v1 sessions CLI tests
+func sessionsV1TestsCLICommand(cmdArgs ...string) bool {
+	// test #1, list sessions
+	cmdOut := runBosCLIList(cmdArgs...)
 	if cmdOut == nil {
 		return false
 	}
 
-	// Validate that at least we can decode the JSON into a list of strings
+	// BOS v1: Validate that at least we can decode the JSON into a list of strings
 	stringList, err := common.DecodeJSONIntoStringList(cmdOut)
 	if err != nil {
 		common.Error(err)
 		return false
 	}
 
-	// test #2, list session with session_id
-	// use results from previous tests, grab the first session
+	// Grab the first session ID
 	if len(stringList) == 0 {
-		common.VerbosePrintDivider()
-		common.Infof("skipping test CLI describe sessiontemplate {session_template_id}")
+		common.Infof("skipping test CLI describe {session_id}")
 		common.Infof("results from previous test is []")
 		return true
 	}
 
 	// A session id is available
 	sessionId := stringList[0]
+	common.Infof("Found BOS v1 session with ID '%s'", sessionId)
 
-	common.Infof("Describing BOS session %s via CLI", sessionId)
-	cmdOut = runCLICommand(vnum, "session", "describe", sessionId)
+	// test #2, describe session with session_id
+	if !basicCLIDescribeVerifyStringMapTest(sessionId, cmdArgs...) {
+		return false
+	}
+
+	return true
+}
+
+// v2 sessions CLI tests
+func sessionsV2TestsCLICommand(cmdArgs ...string) bool {
+	// test #1, list sessions
+	cmdOut := runBosCLIList(cmdArgs...)
 	if cmdOut == nil {
 		return false
 	}
 
-	// Validate that the session can at least be decoded into a string map
-	_, err = common.DecodeJSONIntoStringMap(cmdOut)
+	// BOS v2: Decode JSON into a list
+	sessionList, err := common.DecodeJSONIntoList(cmdOut)
+	if err != nil {
+		common.Error(err)
+		return false
+	} else if len(sessionList) == 0 {
+		common.Infof("skipping test CLI describe {session_id}")
+		common.Infof("results from previous test is []")
+		return true
+	}
+
+	// Take the first session listed. It should be a dictionary object.
+	sessionObject := sessionList[0]
+	session, ok := sessionObject.(map[string]interface{})
+	if !ok {
+		common.Errorf("First BOS session listed is not a dictionary object: %v", sessionObject)
+		return false
+	}
+
+	common.Debugf("Getting 'name' field BOS session: %v", session)
+	sessionId, err := common.GetStringFieldFromMapObject("name", session)
 	if err != nil {
 		common.Error(err)
 		return false
 	}
+	common.Infof("Found BOS v2 session with name '%s'", sessionId)
 
-	// TODO: deeper validation of returned response
+	// test #2, describe session with session_id
+	cmdOut = runBosCLIDescribe(sessionId, cmdArgs...)
+	if cmdOut == nil {
+		return false
+	} else if !ValidateV2Session(cmdOut, sessionId) {
+		return false
+	}
+
 	return true
 }