+++ cmsdev/internal/test/bos/bos_sessiontemplate.go
@@ -33,8 +33,83 @@ package bos
 import (
 	"net/http"
 	"stash.us.cray.com/SCMS/cms-tools/cmsdev/internal/lib/common"
-	"stash.us.cray.com/SCMS/cms-tools/cmsdev/internal/lib/test"
 )
+
+const bosV1SessionTemplatesUri = bosV1BaseUri + "/sessiontemplate"
+const bosV2SessionTemplatesUri = bosV2BaseUri + "/sessiontemplates"
+
+const bosV1SessionTemplateTemplateUri = bosV1BaseUri + "/sessiontemplatetemplate"
+const bosV2SessionTemplateTemplateUri = bosV2BaseUri + "/sessiontemplatetemplate"
+
+const bosV1SessionTemplatesCLI = "sessiontemplate"
+const bosV2SessionTemplatesCLI = "sessiontemplates"
+const bosDefaultSessionTemplatesCLI = bosV2SessionTemplatesCLI
+
+const bosV1SessionTemplateTemplateCLI = "sessiontemplatetemplate"
+const bosV2SessionTemplateTemplateCLI = "sessiontemplatetemplate"
+const bosDefaultSessionTemplateTemplateCLI = bosV2SessionTemplateTemplateCLI
+
+func sessionTemplatesTestsAPI(params *common.Params) (passed bool) {
+	passed = true
+
+	// v1 session template template
+	if !sessionTemplateTemplateTestURI(bosV1SessionTemplateTemplateUri, params) {
+		passed = false
+	}
+
+	// v2 session template template
+	if !sessionTemplateTemplateTestURI(bosV2SessionTemplateTemplateUri, params) {
+		passed = false
+	}
+
+	// v1 session templates
+	if !sessionTemplatesTestsURI(bosV1SessionTemplatesUri, params) {
+		passed = false
+	}
+
+	// v2 session templates
+	if !sessionTemplatesTestsURI(bosV2SessionTemplatesUri, params) {
+		passed = false
+	}
+
+	return
+}
+
+func sessionTemplatesTestsCLI() (passed bool) {
+	passed = true
+
+	// v1 session template template
+	if !sessionTemplateTemplateTestCLICommand("v1", bosV1SessionTemplateTemplateCLI) {
+		passed = false
+	}
+
+	// v2 session template template
+	if !sessionTemplateTemplateTestCLICommand("v2", bosV2SessionTemplateTemplateCLI) {
+		passed = false
+	}
+
+	// default (v2) session template template
+	if !sessionTemplateTemplateTestCLICommand(bosDefaultSessionTemplateTemplateCLI) {
+		passed = false
+	}
+
+	// v1 session templates
+	if !sessionTemplatesTestsCLICommand("v1", bosV1SessionTemplatesCLI) {
+		passed = false
+	}
+
+	// v2 session templates
+	if !sessionTemplatesTestsCLICommand("v2", bosV2SessionTemplatesCLI) {
+		passed = false
+	}
+
+	// default (v2) session templates
+	if !sessionTemplatesTestsCLICommand(bosDefaultSessionTemplatesCLI) {
+		passed = false
+	}
+
+	return
+}
 
 func getFirstSessionTemplateId(listCmdOut []byte) (string, error) {
 	return common.GetStringFieldFromFirstItem("name", listCmdOut)
@@ -49,125 +124,78 @@ func ValidateSessionTemplateId(mapCmdOut []byte, expectedName string) bool {
 	return true
 }
 
-// sessiontemplate tests
-func sessionTemplateTestsAPI() bool {
-	var baseurl string = common.BASEURL
-	const totalNumTests int = 2
+// sessiontemplatetemplate API test
+func sessionTemplateTemplateTestURI(uri string, params *common.Params) bool {
+	return basicGetUriVerifyStringMapTest(uri, params)
+}
 
-	numTests, numTestsFailed := 0, 0
-	params := test.GetAccessTokenParams()
-	if params == nil {
-		return false
-	}
+// sessiontemplatetemplate CLI test
+func sessionTemplateTemplateTestCLICommand(cmdArgs ...string) bool {
+	return basicCLIListVerifyStringMapTest(cmdArgs...)
+}
 
-	// test #1, get session template template
-	url := baseurl + endpoints["bos"]["sessiontemplatetemplate"].Url
-	numTests++
-	test.RestfulTestHeader("GET sessiontemplatetemplate", numTests, totalNumTests)
-	resp, err := test.RestfulVerifyStatus("GET", url, *params, http.StatusOK)
-	if err != nil {
-		common.Error(err)
-		numTestsFailed++
-	} else {
-		// Validate that we can decode it into a map object
-		_, err = common.DecodeJSONIntoStringMap(resp.Body())
-		if err != nil {
-			common.Error(err)
-			numTestsFailed++
-		}
-		// TODO: deeper validation of returned response
-	}
-
-	// test #2, list session templates
-	url = baseurl + endpoints["bos"]["sessiontemplate"].Url
-	numTests++
-	test.RestfulTestHeader("GET sessiontemplate", numTests, totalNumTests)
-	resp, err = test.RestfulVerifyStatus("GET", url, *params, http.StatusOK)
+// session templates API tests
+func sessionTemplatesTestsURI(uri string, params *common.Params) bool {
+	// test #1, list session templates
+	common.Infof("GET %s test scenario", uri)
+	resp, err := bosRestfulVerifyStatus("GET", uri, params, http.StatusOK)
 	if err != nil {
 		common.Error(err)
 		return false
 	}
-	// TODO: deeper validation of returned response
 
-	// test #3, list sessiontemplate with session_template_id
-	// use results from previous tests, grab the first sessiontemplate
+	// use results from previous test, grab the first session template
 	sessionTemplateId, err := getFirstSessionTemplateId(resp.Body())
 	if err != nil {
 		common.Error(err)
 		return false
 	} else if len(sessionTemplateId) == 0 {
-		common.VerbosePrintDivider()
-		common.Infof("skipping test GET /sessiontemplate/{session_template_id}")
+		common.Infof("skipping test GET %s/{session_template_id}", uri)
 		common.Infof("results from previous test is []")
-		return numTestsFailed == 0
+		return true
 	}
 
 	// a session_template_id is available
-	url = baseurl + endpoints["bos"]["sessiontemplate"].Url + "/" + sessionTemplateId
-	numTests++
-	test.RestfulTestHeader("GET session_template_id", numTests, totalNumTests)
-	resp, err = test.RestfulVerifyStatus("GET", url, *params, http.StatusOK)
+	uri += "/" + sessionTemplateId
+	common.Infof("GET %s test scenario", uri)
+	resp, err = bosRestfulVerifyStatus("GET", uri, params, http.StatusOK)
 	if err != nil {
 		common.Error(err)
-		numTestsFailed++
+		return false
 	} else if !ValidateSessionTemplateId(resp.Body(), sessionTemplateId) {
-		numTestsFailed++
+		return false
 	}
 
-	// TODO: deeper validation of returned response
-
-	return numTestsFailed == 0
+	return true
 }
 
-func sessionTemplateTestsCLI(vnum int) bool {
-	var err error
-	var numTestsFailed = 0
-
-	// test #1, get example session template
-	common.Infof("Getting example BOS session template via CLI")
-	cmdOut := runCLICommand(vnum, "sessiontemplatetemplate", "list")
-	if cmdOut == nil {
-		numTestsFailed++
-	} else {
-		// Validate that we can decode it into a map object
-		_, err = common.DecodeJSONIntoStringMap(cmdOut)
-		if err != nil {
-			common.Error(err)
-			numTestsFailed++
-		}
-		// TODO: deeper validation of returned response
-	}
-
-	// test #2, list session templates
-	common.Infof("Getting list of all BOS session templates via CLI")
-	cmdOut = runCLICommand(vnum, "sessiontemplate", "list")
+// session templates CLI tests
+func sessionTemplatesTestsCLICommand(cmdArgs ...string) bool {
+	// test #1, list session templates
+	cmdOut := runBosCLIList(cmdArgs...)
 	if cmdOut == nil {
 		return false
 	}
-	// TODO: deeper validation of returned response
 
-	// test #3, list sessiontemplate with session_template_id
-	// use results from previous tests, grab the first sessiontemplate
+	// use results from previous test, grab the first session template
 	sessionTemplateId, err := getFirstSessionTemplateId(cmdOut)
 	if err != nil {
 		common.Error(err)
 		return false
 	} else if len(sessionTemplateId) == 0 {
-		common.VerbosePrintDivider()
-		common.Infof("skipping test CLI describe sessiontemplate {session_template_id}")
+		common.Infof("skipping test CLI describe component {session_template_id}")
 		common.Infof("results from previous test is []")
-		return numTestsFailed == 0
+		return true
 	}
 
-	// a session_template_id is available
-	common.Infof("Describing BOS session template %s via CLI", sessionTemplateId)
-	cmdOut = runCLICommand(vnum, "sessiontemplate", "describe", sessionTemplateId)
+	// a session template id is available
+	// test #2 describe session templates
+	cmdOut = runBosCLIDescribe(sessionTemplateId, cmdArgs...)
 	if cmdOut == nil {
 		return false
 	} else if !ValidateSessionTemplateId(cmdOut, sessionTemplateId) {
 		return false
 	}
 
-	// TODO: deeper validation of returned response
-	return numTestsFailed == 0
+	return true
 }